Finished the rework. Tests pass, so obviously it's done, right?
