+++ src/layouts/layout_v1.rs
@@ -5,183 +5,187 @@
 
 #![allow(non_snake_case, non_camel_case_types)]
 
-use std::fs;
-use std::io::Write;
-use std::sync::Mutex;
+use serde::{Deserialize, Serialize};
 
-use anyhow::Result;
-use once_cell::sync::Lazy;
+use crate::layouts::shared::NamedAnchor;
+use crate::plugin::{
+    Align, Color, HudElement, LayoutFlattened, Point, SlotFlattened, TextFlattened,
+};
 
-use crate::plugin::{HudLayout1, NamedAnchor, Point};
-
-static LAYOUT_PATH: &str = "./data/SKSE/Plugins/SoulsyHUD_Layout.toml";
-
-/// There can be only one. Not public because we want access managed.
-static LAYOUT: Lazy<Mutex<HudLayout1>> = Lazy::new(|| Mutex::new(HudLayout1::init()));
-
-/// Lazy parsing of the compile-time include of the default layout, as a fallback.
-static DEFAULT_LAYOUT: Lazy<HudLayout1> = Lazy::new(HudLayout1::default);
-
-#[cfg(not(test))]
-use crate::plugin::{resolutionHeight, resolutionWidth};
-
-// mocked screen resolution numbers, because these functions are provided by
-// C++ and require imgui etc.
-#[cfg(test)]
-fn resolutionWidth() -> f32 {
-    3440.0
+#[derive(Deserialize, Serialize, Debug, Clone)]
+pub struct HudLayout1 {
+    #[serde(default)]
+    /// A global scaling factor for the entire hud.
+    global_scale: f32,
+    /// Where to draw the HUD; an offset from the top left corner.
+    #[serde(default)]
+    anchor: Point,
+    #[serde(
+        default,
+        deserialize_with = "crate::layouts::shared::deserialize_named_anchor"
+    )]
+    pub anchor_name: NamedAnchor, // pub for tests
+    /// The dimensions of a bounding box for the HUD.
+    size: Point,
+    /// The color to draw the HUD bg image with; if zero will not be drawn.
+    bg_color: Color,
+    /// Hide the ammo slot if a ranged weapon is not equipped.
+    #[serde(default)]
+    hide_ammo_when_irrelevant: bool,
+    /// Hide the left hand slot when a ranged weapon is equipped.
+    #[serde(default)]
+    hide_left_when_irrelevant: bool,
+    /// One slot layout for each element. This wants to be map, not a vec,
+    /// but the map types are not shareable.
+    layouts: Vec<SlotLayout>,
+    font: String,
+    /// The font size for most things; a hint to the font loader.
+    font_size: f32,
+    /// Whether to buld glyphs for full Chinese text display.
+    #[serde(default)]
+    chinese_full_glyphs: bool,
+    /// Whether to build glyphs for simplified Chinese text display.
+    #[serde(default)]
+    simplified_chinese_glyphs: bool,
+    /// Whether to build glyphs for simplified Chinese text display.
+    #[serde(default)]
+    cyrillic_glyphs: bool,
+    /// Whether to build glyphs for Cyrillic text display.
+    #[serde(default)]
+    japanese_glyphs: bool,
+    /// Whether to build glyphs for Japanese text display.
+    #[serde(default)]
+    korean_glyphs: bool,
+    /// Whether to build glyphs for Thai text display.
+    #[serde(default)]
+    thai_glyphs: bool,
+    /// Whether to build glyphs for Vietnamese text display.
+    #[serde(default)]
+    vietnamese_glyphs: bool,
 }
 
-#[cfg(test)]
-fn resolutionHeight() -> f32 {
-    1440.0
-}
+#[derive(Deserialize, Serialize, Debug, Clone)]
+pub struct SlotLayout {
+    /// The hud element this layout is for.
+    element: HudElement,
+    /// The name of the hud element this layout is for. For humans.
+    name: String,
+    /// How to align any text associated with this slot.
+    #[serde(
+        default,
+        deserialize_with = "crate::layouts::shared::deserialize_align"
+    )]
+    align_text: Align,
+    /// An offset from the overall hud anchor point to draw this element at.
+    offset: Point,
+    /// The size of this element, to scale everything to.
+    size: Point,
+    /// The color of any background for this element. If its alpha is 0, the bg is not drawn.
+    bg_color: Color,
 
-/// Read our layout data from the file, or fall back to defaults if the file
-/// is not present or is invalid TOML.
-pub fn hud_layout() -> HudLayout1 {
-    let layout = LAYOUT
-        .lock()
-        .expect("Unrecoverable runtime problem: cannot acquire layout lock.");
-    layout.clone()
+    /// The color of any icon for this element. If its alpha is 0, the icon is not drawn.
+    icon_color: Color,
+    /// The size of the icon to draw in this slot.
+    icon_size: Point,
+    /// Where to draw the icon; a center point relative to the center of this slot.
+    #[serde(default)]
+    icon_offset: Point,
+
+    /// The color to use for this element's hotkey, if it has one. If alpha is zero, it's not drawn.
+    hotkey_color: Color,
+    /// Where to draw this hotkey, relative to the anchor point.
+    hotkey_offset: Point,
+    /// Scale for any hotkey icon.
+    hotkey_size: Point,
+    /// The color to use to draw the key. Not drawn if the alpha is zero.
+    hotkey_bg_color: Color,
+
+    /// If text is drawn in this element, where to draw it.
+    count_offset: Point,
+    /// If this element has to show a count, the font size to use.
+    count_font_size: f32,
+    /// The color of any count size text; 0 alpha means not to draw it at all.
+    count_color: Color,
+
+    /// The color of any item name text; 0 alpha means not to draw it at all.
+    name_color: Color,
+    /// Where to draw the item name.
+    name_offset: Point,
+    /// The font size to use for this item's name.
+    #[serde(default)]
+    name_font_size: f32,
 }
 
 impl HudLayout1 {
-    /// Read a layout object from a toml file.
-    pub fn read_from_file(pathstr: &str) -> Result<Self> {
-        let path = std::path::Path::new(pathstr);
-        if !path.exists() {
-            // No file? We write out defaults.
-            let layout = DEFAULT_LAYOUT.clone();
-            let buf = toml::to_string_pretty(&layout)?;
-            let mut fp = fs::File::create(path)?;
-            write!(fp, "{buf}")?;
-            Ok(layout)
-        } else if let Ok(buf) = fs::read_to_string(path) {
-            match toml::from_str::<HudLayout1>(&buf) {
-                Ok(v) => Ok(v),
-                Err(e) => {
-                    // We are *not* overwriting a bad TOML file, but we are logging it.
-                    // The player might be editing it and experimenting.
-                    log::warn!("Bad TOML in hud layout!; {e:?}");
-                    Ok(DEFAULT_LAYOUT.clone())
-                }
-            }
-        } else {
-            log::warn!(
-                "Unable to read any data from {}! Falling back to defaults",
-                LAYOUT_PATH
-            );
-            Ok(DEFAULT_LAYOUT.clone())
-        }
-    }
-
-    /// Refresh the layout from the file, to take an out-of-band update and apply it in-game.
-    pub fn refresh() {
-        let old = hud_layout();
-        if let Ok(buf) = toml::to_string_pretty(&old) {
-            let backup = format!("{LAYOUT_PATH}.bak");
-            let path = std::path::Path::new(&backup);
-            if let Ok(mut fp) = fs::File::create(path) {
-                if write!(fp, "{buf}").is_ok() {
-                    log::info!(
-                        "Previous layout file has been backed up to {}",
-                        path.display()
-                    );
-                }
-            }
-        }
-
-        match HudLayout1::read_from_file(LAYOUT_PATH) {
-            Ok(v) => {
-                log::info!(
-                    "hud layout read: loc={}, {}; size={}, {}; global scale={};",
-                    v.anchor_point().x,
-                    v.anchor_point().y,
-                    v.size.x,
-                    v.size.y,
-                    v.global_scale
-                );
-                let mut hudl = LAYOUT
-                    .lock()
-                    .expect("Unrecoverable runtime problem: cannot acquire layout lock.");
-                *hudl = v;
-            }
-            Err(e) => {
-                log::warn!("Failed to read layout file; continuing with previous; {e:?}");
-            }
-        }
-    }
-
-    pub fn init() -> HudLayout1 {
-        match HudLayout1::read_from_file(LAYOUT_PATH) {
-            Ok(v) => {
-                log::info!("Successfully initialized HUD layout from TOML file.");
-                v
-            }
-            Err(e) => {
-                log::warn!("Failed to read TOML layout file; initializing from defaults; {e:?}");
-                DEFAULT_LAYOUT.clone()
-            }
-        }
-    }
-
     pub fn anchor_point(&self) -> Point {
-        // If we read a named anchor point, turn it into pixels.
-        // The anchor point is the location of the hud CENTER, so we offset.
-        let screen_width = resolutionWidth();
-        let screen_height = resolutionHeight();
-        let width = self.size.x * self.global_scale;
-        let height = self.size.y * self.global_scale;
+        super::anchor_point(
+            self.global_scale,
+            &self.size,
+            &self.anchor_name,
+            Some(&self.anchor),
+        )
+    }
+}
 
-        match self.anchor_name {
-            NamedAnchor::TopLeft => Point {
-                x: width / 2.0,
-                y: height / 2.0,
+impl From<&HudLayout1> for LayoutFlattened {
+    fn from(v: &HudLayout1) -> Self {
+        let slots = v.layouts.iter().map(SlotFlattened::from).collect();
+
+        LayoutFlattened {
+            global_scale: v.global_scale,
+            anchor: v.anchor_point(),
+            bg_size: v.size.clone(),
+            bg_color: v.bg_color.clone(),
+            bg_image: "hud_bg.svg".to_string(),
+            hide_ammo_when_irrelevant: v.hide_ammo_when_irrelevant,
+            hide_left_when_irrelevant: v.hide_left_when_irrelevant,
+            font: v.font.clone(),
+            font_size: v.font_size,
+            chinese_full_glyphs: v.chinese_full_glyphs,
+            simplified_chinese_glyphs: v.simplified_chinese_glyphs,
+            cyrillic_glyphs: v.cyrillic_glyphs,
+            japanese_glyphs: v.japanese_glyphs,
+            korean_glyphs: v.korean_glyphs,
+            thai_glyphs: v.thai_glyphs,
+            vietnamese_glyphs: v.vietnamese_glyphs,
+            slots,
+        }
+    }
+}
+
+impl From<&SlotLayout> for SlotFlattened {
+    fn from(xs: &SlotLayout) -> Self {
+        let text = vec![
+            TextFlattened {
+                offset: xs.name_offset.clone(),
+                color: xs.name_color.clone(),
+                alignment: xs.align_text,
+                contents: "{name}".to_string(),
+                font_size: xs.name_font_size,
             },
-            NamedAnchor::TopRight => Point {
-                x: screen_width - width / 2.0,
-                y: height / 2.0,
+            TextFlattened {
+                offset: xs.count_offset.clone(),
+                color: xs.count_color.clone(),
+                alignment: xs.align_text,
+                contents: "{count}".to_string(),
+                font_size: xs.count_font_size,
             },
-            NamedAnchor::BottomLeft => Point {
-                x: width / 2.0,
-                y: screen_height - height / 2.0,
-            },
-            NamedAnchor::BottomRight => Point {
-                x: screen_width - width / 2.0,
-                y: screen_height - height / 2.0,
-            },
-            NamedAnchor::Center => Point {
-                x: screen_width / 2.0,
-                y: screen_height / 2.0,
-            },
-            NamedAnchor::CenterTop => Point {
-                x: screen_width / 2.0,
-                y: height / 2.0,
-            },
-            NamedAnchor::CenterBottom => Point {
-                x: screen_width / 2.0,
-                y: screen_height - height / 2.0,
-            },
-            NamedAnchor::LeftCenter => Point {
-                x: width / 2.0,
-                y: screen_height / 2.0,
-            },
-            NamedAnchor::RightCenter => Point {
-                x: screen_width - width / 2.0,
-                y: screen_height / 2.0,
-            },
-            _ => {
-                if self.anchor == Point::default() {
-                    log::info!("Layout has neither a named anchor nor an anchor point. Falling back to top left.");
-                    Point {
-                        x: width / 2.0,
-                        y: height / 2.0,
-                    }
-                } else {
-                    self.anchor.clone()
-                }
-            }
+        ];
+        SlotFlattened {
+            element: xs.element,
+            bg_size: xs.size.clone(),
+            bg_color: xs.bg_color.clone(),
+            bg_image: "slot_bg.svg".to_string(),
+            icon_size: xs.icon_size.clone(),
+            icon_offset: xs.icon_offset.clone(),
+            icon_color: xs.icon_color.clone(),
+            hotkey_size: xs.hotkey_size.clone(),
+            hotkey_offset: xs.hotkey_offset.clone(),
+            hotkey_color: xs.hotkey_color.clone(),
+            hotkey_bg_size: xs.hotkey_size.clone(),
+            hotkey_bg_color: xs.hotkey_bg_color.clone(),
+            hotkey_bg_image: "key_bg.svg".to_string(),
+            text,
         }
     }
 }
