+++ src/layouts/layout_v2.rs
@@ -1,7 +1,9 @@
 use serde::{Deserialize, Serialize};
 
 use super::shared::*;
-use crate::plugin::{resolutionHeight, resolutionWidth, Align, Color, NamedAnchor, Point};
+use crate::plugin::{
+    Align, Color, HudElement, LayoutFlattened, Point, SlotFlattened, TextFlattened,
+};
 
 /// Where to arrange the HUD elements and what color to draw them in.
 ///
@@ -10,50 +12,50 @@ use crate::plugin::{resolutionHeight, resolutionWidth, Align, Color, NamedAnchor
 pub struct HudLayout2 {
     #[serde(default)]
     /// A global scaling factor for the entire hud.
-    pub global_scale: f32,
+    global_scale: f32,
     /// Where to draw the HUD; an offset from the top left corner.
     #[serde(default)]
-    pub anchor: Point,
+    anchor: Point,
     #[serde(default, deserialize_with = "deserialize_named_anchor")]
-    pub anchor_name: NamedAnchor,
-    pub background: ImageElement,
-    pub right: SlotElement,
-    pub left: SlotElement,
-    pub power: SlotElement,
-    pub utility: SlotElement,
-    pub ammo: SlotElement,
-    pub equipset: SlotElement,
+    anchor_name: NamedAnchor,
+    background: ImageElement,
+    right: SlotElement,
+    left: SlotElement,
+    power: SlotElement,
+    utility: SlotElement,
+    ammo: SlotElement,
+    equipset: SlotElement,
     /// Hide the ammo slot if a ranged weapon is not equipped.
     #[serde(default)]
-    pub hide_ammo_when_irrelevant: bool,
+    hide_ammo_when_irrelevant: bool,
     /// Hide the left hand slot when a ranged weapon is equipped.
     #[serde(default)]
-    pub hide_left_when_irrelevant: bool,
+    hide_left_when_irrelevant: bool,
     /// truetype file to load
-    pub font: String,
+    font: String,
     /// The font size for most things; a hint to the font loader.
-    pub font_size: f32,
+    font_size: f32,
     /// Whether to buld glyphs for full Chinese text display.
     #[serde(default)]
-    pub chinese_full_glyphs: bool,
+    chinese_full_glyphs: bool,
     /// Whether to build glyphs for simplified Chinese text display.
     #[serde(default)]
-    pub simplified_chinese_glyphs: bool,
+    simplified_chinese_glyphs: bool,
     /// Whether to build glyphs for simplified Chinese text display.
     #[serde(default)]
-    pub cyrillic_glyphs: bool,
+    cyrillic_glyphs: bool,
     /// Whether to build glyphs for Cyrillic text display.
     #[serde(default)]
-    pub japanese_glyphs: bool,
+    japanese_glyphs: bool,
     /// Whether to build glyphs for Japanese text display.
     #[serde(default)]
-    pub korean_glyphs: bool,
+    korean_glyphs: bool,
     /// Whether to build glyphs for Thai text display.
     #[serde(default)]
-    pub thai_glyphs: bool,
+    thai_glyphs: bool,
     /// Whether to build glyphs for Vietnamese text display.
     #[serde(default)]
-    pub vietnamese_glyphs: bool,
+    vietnamese_glyphs: bool,
 }
 
 #[derive(Deserialize, Serialize, Debug, Clone, Default)]
@@ -72,62 +74,12 @@ impl HudLayout2 {
     }
 
     pub fn anchor_point(&self) -> Point {
-        // If we read a named anchor point, turn it into pixels.
-        // The anchor point is the location of the hud CENTER, so we offset.
-        let screen_width = resolutionWidth();
-        let screen_height = resolutionHeight();
-        let width = self.background.size.x * self.global_scale;
-        let height = self.background.size.y * self.global_scale;
-
-        match self.anchor_name {
-            NamedAnchor::TopLeft => Point {
-                x: width / 2.0,
-                y: height / 2.0,
-            },
-            NamedAnchor::TopRight => Point {
-                x: screen_width - width / 2.0,
-                y: height / 2.0,
-            },
-            NamedAnchor::BottomLeft => Point {
-                x: width / 2.0,
-                y: screen_height - height / 2.0,
-            },
-            NamedAnchor::BottomRight => Point {
-                x: screen_width - width / 2.0,
-                y: screen_height - height / 2.0,
-            },
-            NamedAnchor::Center => Point {
-                x: screen_width / 2.0,
-                y: screen_height / 2.0,
-            },
-            NamedAnchor::CenterTop => Point {
-                x: screen_width / 2.0,
-                y: height / 2.0,
-            },
-            NamedAnchor::CenterBottom => Point {
-                x: screen_width / 2.0,
-                y: screen_height - height / 2.0,
-            },
-            NamedAnchor::LeftCenter => Point {
-                x: width / 2.0,
-                y: screen_height / 2.0,
-            },
-            NamedAnchor::RightCenter => Point {
-                x: screen_width - width / 2.0,
-                y: screen_height / 2.0,
-            },
-            _ => {
-                if self.anchor == Point::default() {
-                    log::info!("Layout has neither a named anchor nor an anchor point. Falling back to top left.");
-                    Point {
-                        x: width / 2.0,
-                        y: height / 2.0,
-                    }
-                } else {
-                    self.anchor.clone()
-                }
-            }
-        }
+        super::anchor_point(
+            self.global_scale,
+            &self.background.size,
+            &self.anchor_name,
+            None,
+        )
     }
 }
 
@@ -199,4 +151,61 @@ pub struct TextElement {
 pub struct ProgressElement {
     offset: Point,
     color: Color,
+}
+
+impl From<&HudLayout2> for LayoutFlattened {
+    fn from(v: &HudLayout2) -> Self {
+        let slots = vec![
+            flatten(&v.power, HudElement::Power),
+            flatten(&v.utility, HudElement::Utility),
+            flatten(&v.left, HudElement::Left),
+            flatten(&v.right, HudElement::Right),
+            flatten(&v.ammo, HudElement::Ammo),
+            flatten(&v.equipset, HudElement::EquipSet),
+        ];
+
+        LayoutFlattened {
+            global_scale: v.global_scale,
+            anchor: v.anchor_point(),
+            bg_size: v.background.size.clone(),
+            bg_color: v.background.color.clone(),
+            bg_image: v.background.svg.clone(),
+            hide_ammo_when_irrelevant: v.hide_ammo_when_irrelevant,
+            hide_left_when_irrelevant: v.hide_left_when_irrelevant,
+            font: v.font.clone(),
+            font_size: v.font_size,
+            chinese_full_glyphs: v.chinese_full_glyphs,
+            simplified_chinese_glyphs: v.simplified_chinese_glyphs,
+            cyrillic_glyphs: v.cyrillic_glyphs,
+            japanese_glyphs: v.japanese_glyphs,
+            korean_glyphs: v.korean_glyphs,
+            thai_glyphs: v.thai_glyphs,
+            vietnamese_glyphs: v.vietnamese_glyphs,
+            slots,
+        }
+    }
+}
+
+fn flatten(slot: &SlotElement, element: HudElement) -> SlotFlattened {
+    let bg = slot.background.clone().unwrap_or_default();
+    let hotkey = slot.hotkey.clone().unwrap_or_default();
+    let hkbg = hotkey.background.unwrap_or_default();
+    let text = slot.text.iter().map(TextFlattened::from).collect();
+
+    SlotFlattened {
+        element,
+        bg_size: bg.size,
+        bg_color: bg.color,
+        bg_image: bg.svg,
+        icon_size: slot.icon.size.clone(),
+        icon_offset: slot.icon.offset.clone(),
+        icon_color: slot.icon.color.clone(),
+        hotkey_size: hotkey.size,
+        hotkey_offset: hotkey.offset,
+        hotkey_color: hotkey.color,
+        hotkey_bg_size: hkbg.size,
+        hotkey_bg_color: hkbg.color,
+        hotkey_bg_image: hkbg.svg,
+        text,
+    }
 }
